Squash adam/performance-debug changes
Squashed commit of the following:

commit 1ccc21a0a9f268e1fcdc57108c5e095e08077166
Merge: 28cd999a15 3d73d71c18
Author: Adam Shelly <ashelly@anki.com>
Date:   Fri Nov 10 16:10:33 2017 -0800

    Merge branch 'master' into adam/performance-debug

commit 28cd999a15b37696d6b28cd55fd36b408680bad6
Author: Adam Shelly <ashelly@anki.com>
Date:   Fri Nov 10 16:04:21 2017 -0800

    spine refactor fix clock issues

commit 9a2a4237808b1c2d0d369414057df4ca34957605
Author: Adam Shelly <ashelly@anki.com>
Date:   Fri Nov 10 13:00:48 2017 -0800

    experiments with select and sleep

commit a895af423729a0ab322a06b72646aa886092d4fc
Author: Adam Shelly <ashelly@anki.com>
Date:   Thu Nov 9 18:15:55 2017 -0800

    spine serial refactor

commit 11c7734d58f048e797a7ef8bdc372c36e310b164
Author: Adam Shelly <ashelly@anki.com>
Date:   Thu Nov 2 19:28:57 2017 -0700

    remove extra spine debug

commit e39afa28f0748f72fa406253f2c0b2bbccdd9cd3
Author: Adam Shelly <ashelly@anki.com>
Date:   Thu Nov 2 19:26:06 2017 -0700

    fix spine resync algorithm

+++ robot2/core/inc/core/clock.h
@@ -3,9 +3,9 @@
 
 /************* CLOCK Interface ***************/
 
-#define NSEC_PER_SEC  1000000000
-#define NSEC_PER_MSEC 1000000
-#define NSEC_PER_USEC 1000
+#define NSEC_PER_SEC  ((uint64_t)1000000000)
+#define NSEC_PER_MSEC ((uint64_t)1000000)
+#define NSEC_PER_USEC ((uint64_t)1000)
 
 
 uint64_t steady_clock_now(void);
